simplified resource loading for speed at the cost of consistency
once in a while it wont load entirely correctly, but it will load a lot faster now.
Just hit show town, when it happens to not show.
This way we avoid full page freeze, which is more annoying. There is probably some better way to handle this, that I haven't found yet. for covering best of both worlds.